ai-backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
pydantic>=2.7.1
python-multipart>=0.0.9
openai-whisper>=20231117
ollama>=0.2.1
aiofiles>=23.2.1
